backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
python-multipart==0.0.17
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0
alembic==1.14.0

# Vector Database
qdrant-client==1.12.1

# AI and ML
openai==1.55.0
anthropic==0.39.0
google-generativeai==0.8.3
tiktoken==0.8.0

# Authentication
pyjwt==2.10.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.28.0
aiohttp==3.11.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# CORS
python-cors==1.0.0
